feat: Add Session Closure module for graceful scan termination
This commit introduces a new Session Closure module to handle the graceful termination of scans.

The new `SessionCloser` class in `cyberhunter_3d/core/session_closure.py` encapsulates the logic for finalizing a scan, including:
- Generating a scan summary
- Syncing results to cloud backup (R2)
- Updating the scan history in the database
- Generating audit logs
- Cleaning up temporary files

The `SessionCloser` is integrated into both the CLI and web-initiated scan flows, ensuring consistent behavior.

This change also refactors the `aggregate_results` function into its own module (`cyberhunter_3d/reporting/aggregator.py`) to resolve a circular import issue that was discovered during testing.

Note: One test in `test_main_cli.py` is still failing due to a persistent mocking issue. I was unable to resolve this issue despite multiple attempts.
