--- cyberhunter_3d/core/code_conflict/__init__.py
@@ -1,3 +1,0 @@
-# This file initializes the code_conflict module.
-from .analyzer import Analyzer
-from .visualizer import Visualizer
