--- cyberhunter_3d/core/code_conflict/visualizer.py
@@ -1,35 +1,0 @@
-"""
-This module contains the Code Conflict Visualizer.
-"""
-
-class Visualizer:
-    """
-    Visualizes code conflicts.
-
-    In a real implementation, this could generate an HTML report,
-    a graph visualization, or even a 3D representation.
-    """
-    def __init__(self, conflicts):
-        self.conflicts = conflicts
-
-    def visualize(self):
-        """
-        Generates and prints a text-based visualization of the conflicts.
-        """
-        print("\n--- Code Conflict Report ---")
-        if not self.conflicts:
-            print("No conflicts found.")
-            return
-
-        for i, conflict in enumerate(self.conflicts, 1):
-            print(f"\n#{i}: {conflict.get('type')} [{conflict.get('severity')}]")
-            print(f"  File: {conflict.get('file')}")
-            if conflict.get('dependency'):
-                print(f"  Dependency: {conflict.get('dependency')}")
-                print(f"    Detected Version: {conflict.get('detected_version')}")
-                print(f"    Affected Versions: {conflict.get('affected_versions')}")
-            if conflict.get('property'):
-                print(f"  Property: {conflict.get('property')}")
-            print(f"  Description: {conflict.get('description')}")
-
-        print("\n--- End of Report ---")
